requirements.txt
# LUNA16 Lung Cancer Detection - Dependencies
# Python 3.8+

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Medical Imaging
SimpleITK>=2.2.0
pydicom>=2.3.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Progress Bars
tqdm>=4.65.0

# Utilities
tabulate>=0.9.0
Pillow>=9.5.0

# Jupyter (for notebooks)
jupyter>=1.0.0
notebook>=6.5.0
ipykernel>=6.22.0
ipywidgets>=8.0.0

# Optional but recommended
scipy>=1.10.0
opencv-python>=4.7.0
